WIP: refactoring of vars processing
