+++ internal/conditions/cerbos_lib_test.go
@@ -190,162 +190,6 @@ func TestCmpSelectAndCall(t *testing.T) {
 	v := reflect.DeepEqual(*ast0, *ast)
 	is.True(v)
 }
-func updateIds(e *expr.Expr, n *int64) {
-	if e == nil {
-		return
-	}
-	*n++
-	e.Id = *n
-	switch e := e.ExprKind.(type) {
-	case *expr.Expr_SelectExpr:
-		updateIds(e.SelectExpr.Operand, n)
-	case *expr.Expr_CallExpr:
-		updateIds(e.CallExpr.Target, n)
-		for _, arg := range e.CallExpr.Args {
-			updateIds(arg, n)
-		}
-	case *expr.Expr_StructExpr:
-		for _, entry := range e.StructExpr.Entries {
-			updateIds(entry.GetMapKey(), n)
-			updateIds(entry.GetValue(), n)
-		}
-	case *expr.Expr_ComprehensionExpr:
-		ce := e.ComprehensionExpr
-		updateIds(ce.IterRange, n)
-		updateIds(ce.AccuInit, n)
-		updateIds(ce.LoopStep, n)
-		updateIds(ce.LoopCondition, n)
-		updateIds(ce.Result, n)
-	case *expr.Expr_ListExpr:
-		for _, element := range e.ListExpr.Elements {
-			updateIds(element, n)
-		}
-	}
-}
-
-func replace(e *expr.Expr, vars map[string]*expr.Expr) *expr.Expr {
-	var r func(e *expr.Expr) *expr.Expr
-	r = func(e *expr.Expr) *expr.Expr {
-		if e == nil {
-			return nil
-		}
-		switch e := e.ExprKind.(type) {
-		case *expr.Expr_SelectExpr:
-			ident := e.SelectExpr.Operand.GetIdentExpr()
-			if ident != nil {
-				if ident.Name == conditions.CELVariablesAbbrev || ident.Name == conditions.CELVariablesIdent {
-					if v, ok := vars[e.SelectExpr.Field]; ok {
-						return v
-					} else {
-						panic("unknown variable")
-					}
-				}
-			} else {
-				e.SelectExpr.Operand = r(e.SelectExpr.Operand)
-			}
-		case *expr.Expr_CallExpr:
-			e.CallExpr.Target = r(e.CallExpr.Target)
-			for i, arg := range e.CallExpr.Args {
-				e.CallExpr.Args[i] = r(arg)
-			}
-		case *expr.Expr_StructExpr:
-			for _, entry := range e.StructExpr.Entries {
-				if k, ok := entry.KeyKind.(*expr.Expr_CreateStruct_Entry_MapKey); ok {
-					k.MapKey = r(k.MapKey)
-				}
-				entry.Value = r(entry.Value)
-			}
-		case *expr.Expr_ComprehensionExpr:
-			ce := e.ComprehensionExpr
-			ce.IterRange = r(ce.IterRange)
-			ce.AccuInit = r(ce.AccuInit)
-			ce.LoopStep = r(ce.LoopStep)
-			ce.LoopCondition = r(ce.LoopCondition)
-			// ce.Result seems to be always an identifier, so isn't necessary to process
-		case *expr.Expr_ListExpr:
-			for i, element := range e.ListExpr.Elements {
-				e.ListExpr.Elements[i] = r(element)
-			}
-		}
-		return e
-	}
-
-	var n int64
-	e = r(e)
-	updateIds(e, &n)
-	return e
-}
-
-func TestPartialEvaluationWithGlobalVars(t *testing.T) {
-	is := require.New(t)
-
-	env, err := conditions.StdPartialEnv.Extend(cel.Declarations(
-		decls.NewVar("gb_us", decls.NewListType(decls.String)),
-		decls.NewVar("gbLoc", decls.String),
-		decls.NewVar("ca", decls.String),
-		decls.NewVar("gb", decls.NewListType(decls.String)),
-	))
-	is.NoError(err)
-
-	pvars, _ := cel.PartialVars(map[string]interface{}{
-		"gbLoc": "en_GB",
-		"gb_us": []string{"GB", "US"},
-		"ca":    "ca",
-	}, cel.AttributePattern("R"))
-
-	variables := make(map[string]*expr.Expr)
-	for k, txt := range map[string]string{
-		"locale": `R.attr.language + "_" + R.attr.country`,
-		"geo":    "R.attr.geo",
-		"gb_us":  `["gb", "us"].map(t, t.upperAscii())`,
-		"info":   `{"country": "GB", "language": "en"}`,
-	} {
-		e, iss := env.Compile(txt)
-		is.Nil(iss, iss.Err())
-		variables[k] = e.Expr()
-	}
-	tests := []struct {
-		expr, want string
-	}{
-		{
-			expr: "V.locale == gbLoc",
-			want: `R.attr.language + "_" + R.attr.country == "en_GB"`,
-		},
-		{
-			expr: "V.geo in (gb_us + [ca]).map(t, t.upperAscii())",
-			want: `R.attr.geo in ["GB", "US", "CA"]`,
-		},
-		{
-			expr: "V.geo in (variables.gb_us + [ca]).map(t, t.upperAscii())",
-			want: `R.attr.geo in ["GB", "US", "CA"]`,
-		},
-		{
-			expr: `V.info.language + "_" + V.info.country == gbLoc`,
-			want: "true",
-		},
-	}
-	for _, tt := range tests {
-		t.Run(tt.expr, func(t *testing.T) {
-			is := require.New(t)
-			ast, iss := env.Compile(tt.expr)
-			is.Nil(iss, iss.Err())
-			e := replace(ast.Expr(), variables)
-			ast = cel.ParsedExprToAst(&expr.ParsedExpr{Expr: e})
-			prg, err := env.Program(ast, cel.EvalOptions(cel.OptTrackState, cel.OptPartialEval))
-			is.NoError(err)
-			out, det, err := prg.Eval(pvars)
-			log.Print(types.IsUnknown(out))
-			is.NoError(err)
-
-			residual, err := env.ResidualAst(ast, det)
-			is.NoError(err)
-			astToString, err := cel.AstToString(residual)
-			is.NoError(err)
-			is.Equal(tt.want, astToString)
-		})
-	}
-}
-
 func TestPartialEvaluationWithMacroGlobalVars(t *testing.T) {
 	expander := func(eh parser.ExprHelper, t *expr.Expr, args []*expr.Expr) (*expr.Expr, *common.Error) {
 		return eh.Select(eh.Select(eh.Ident("R"), "attr"), "geo"), nil
